BBox: Move label up by font height / 6
This is for BeOS application compatibility. BeOS apps expect the label
to be drawn a few pixels higher and may not draw correctly otherwise.

Font height / 6 scales with font size so that the label moves up
proportionally at larger font sizes e.g. 21pt. 1/6 is a better offset
for ArtPaint Global settings and DVDRip 1.0.

Change-Id: I6d852111d75a97f34539f6e94c2e2408b6c15808
Reviewed-on: https://review.haiku-os.org/c/haiku/+/3017
Reviewed-by: John Scipione <jscipione@gmail.com>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>