Tracker: list item height must be an integer.
The item height is rounded to the nearest integer when drawing, but it
is also used to compute an item position:
position in pixels = item number * height

If the height is not an integer, this leads to a rounding error, and
depending on how the items are drawn, some of them are shifted up by a
pixel because of it.

The fix is to round up the item height when it is first computed, and
then make all computations keeping integer values.

Fixes #16377.